Move demo controller to dedicated class
